backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
pydantic>=2.11,<3.0
pydantic-settings>=2.6,<3.0
httpx>=0.28,<1.0
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
google-api-python-client>=2.170,<3.0
google-auth>=2.40,<3.0
pytest>=8.4,<9.0
pytest-asyncio>=1.2,<2.0
PyJWT>=2.10,<3.0
psycopg[binary]>=3.2,<4.0
